requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
Django==3.2
django-debug-toolbar==3.2.3
django-model-utils==4.3.1
djangorestframework==3.12.0
exceptiongroup==1.1.1
flake8==6.0.0
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==6.2.5
pytest-django==4.5.2
pytest-pythonpath==0.7.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.6.3
